refactor: use useShell hook for UpdaterBanner
