+++ src/common/Platform/Platform.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext } from 'react';
 import { WHITELISTED_HOSTS } from 'stremio/common/CONSTANTS';
-import useShell from './useShell';
+import { useShell } from 'stremio/common';
 import { name, isMobile } from './device';
 
 interface PlatformContext {
